[ALPS04769014] DISP: porting display kernel-4.14
[Detail]: porting display kernel-4.14

MTK-Commit-Id: f8d780d5b335a5b2ae15fd05c3bcc1ed3737a877

CR-Id: ALPS04769014
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6765/videox/debug.h
@@ -33,7 +33,7 @@ extern int disp_layer_info_statistic(
 	struct disp_ddp_path_config *last_config, struct disp_frame_cfg_t *cfg);
 
 /* show hrt */
-extern int hrt_high, hrt_low;
+extern unsigned long long hrt_high, hrt_low;
 extern int hrt_show_flag;
 
 #ifdef MTKFB_DBG
